Show error if unregistering fails
Include an option to continue anyway. This is to prevent deadlocks.
For example consider this case:

1. Unregister reaches server, server deletes account.
2. HTTP 200 response is dropped, phone shows network error.
3. Phone retries, server now responds with 400 (bad request, account does not exist anymore).

In this case, it would be safe to continue.

And if the user "continues anyway" but the account was not deleted,
the user can still log in again (*) and retry deleting the account.

(*) if this is during the crypto migration (and not from the ServerSettings)
    then the user would need to downgrade the app to log in again.