FMD Server: default to empty URL
Reason: privacy.

When the user opens the AddAccountActivity,
a request is made to <url>/version to obtain the server version.
When the <url> defaults to fmd.nulide.de, this leaks
to Nulide the IPs of everyone who has freshly installed FMD.

To solve this, default to an empty URL.
Users need to opt-in to using Nulide's instance by pressing a button.