Drop fragment from openstreetmap.org URL
In an openstreetmap.org URL, if mlat and mlon query parameters are used,
it's unnecessary to repeat their values in the fragment. The zoom can be
provided as another query parameter, so the fragment can be removed.

If the author remembers correctly, providing the fragment used to reset
user's chosen layer and overlays, since they're stored in the fragment;
though currently visiting any URL with mlat and mlon seem to reset them.