Include a geo URI in locate command response
A geo URI (RFC 5870) makes it easy to open the location in an
interactive maps/navigation app, which the user may prefer over opening
openstreetmap.org in a web browser. The openstreetmap.org link is kept
for devices that do not support geo URIs, but have (for example)
internet access.

Fixes: https://gitlab.com/Nulide/findmydevice/-/issues/200