Remove IntroductionActivity and Permission.java
The introduction was only there to walk through all permissions.
With the new home screen design, the permissions are displayed directly where
they are needed, e.g. in the commands list and in the transport channels list.

No introduction means the user can directly start using FMD, rather than
having to walk through 10+ screens and grant permissions for features they
don't yet understand.

TODO: With this change, the POST_NOTIFICATIONS and IGNORE_BATTERY_PERMISSIONS
are no longer granted. These are global to the app, and not tied to any command
or transport. We will add a separate display mechanism for such global permissions.