Fix: only register UnifiedPush if FMD Server account exists
Also cancel upload job again just to be safe.

Before this change, if the user:

1. Registered an account with FMD Server, and then
2. Deleted the account again.

The PushReceiver.onNewEndpoint would try to register with the server.
But it would fail with HTTP 400, since the server doesn't know any empty account ID.

Other than being spammy, these requests are a privacy issue.