Use locationManager.getCurrentLocation() on Android 12+
Instead of locationManager.requestLocationUpdates().

Testing on my device without GApps shows that getCurrentLocation() more
reliably obtains a _new_ location. At least when you are outside and a GPS
fix can be obtained.

requestLocationUpdates() on the other hand often returned a recent but still
outdated location. This is likely because it returns the last cached location
immediately (if it is fresh enough), and an updated location later.
But our code only returns the first location and then exits.

So for Android 12+ it is easiest to just call getCurrentLocation().